Remove old legacy configs
