--- score/launch_manager/src/daemon/src/alive_monitor/config/hm_flatcfg.fbs
@@ -1,68 +1,0 @@
-namespace HMFlatBuffer;
-
-// HM version 8.1
-
-file_identifier "BHMT";
-file_extension "bin";
-
-enum ProcessType : byte {
-  REGULAR_PROCESS = 0, // REGULAR_PROCESS
-  LM_PROCESS = 1, // LM_PROCESS
-  STM_PROCESS = 2  // STM_PROCESS
-}
-
-table HMEcuCfg {
-  versionMajor: int (id:0);
-  versionMinor: int (id:1);
-  process: [Process] (id:2);
-  hmMonitorInterface: [HmMonitorInterface] (id:3);
-  hmSupervisionCheckpoint: [HmSupervisionCheckpoint] (id:4);
-  hmAliveSupervision: [HmAliveSupervision] (id:5);
-}
-
-table Process {
-  shortName: string (id:0);
-  identifier: string (id:1);
-  index: uint32 (id:2);
-  processType: ProcessType (id:3);
-}
-
-table HmRefProcess {
-  index: uint32 (id:0);
-}
-
-table HmMonitorInterface {
-  instanceSpecifier: string (id:0);
-  processShortName: string (id:1);
-  portPrototype: string (id:2);
-  interfacePath: string (id:3);
-  refProcessIndex: uint32 (id:4);
-  permittedUid: int32 (id:5);
-}
-
-table HmSupervisionCheckpoint {
-  shortName: string (id:0);
-  checkpointId: uint32 (id:1);
-  refInterfaceIndex: uint32 (id:2);
-}
-
-table HmCheckpointTransition {
-  refSourceCPIndex: uint32 (id:0);
-  refTargetCPIndex: uint32 (id:1);
-  infoSource: HmSupervisionCheckpoint (id:2);
-  infoTarget: HmSupervisionCheckpoint (id:3);
-}
-
-table HmAliveSupervision {
-  ruleContextKey: string (id:0);
-  refCheckPointIndex: uint32 (id:1);
-  aliveReferenceCycle: double (id:2);
-  minAliveIndications: uint32 (id:3);
-  maxAliveIndications: uint32 (id:4);
-  isMinCheckDisabled: bool (id:5);
-  isMaxCheckDisabled: bool (id:6);
-  failedSupervisionCyclesTolerance: uint32 (id:7);
-  refProcessIndex: uint32 (id:8);
-}
-
-root_type HMEcuCfg;
