--- score/launch_manager/src/daemon/src/alive_monitor/config/hmcore_flatcfg.fbs
@@ -1,31 +1,0 @@
-namespace HMCOREFlatBuffer;
-
-// HMCORE version 3.0
-
-file_identifier "BHMR";
-file_extension "bin";
-
-table HMCOREEcuCfg {
-  versionMajor: int (id:0);
-  versionMinor: int (id:1);
-  watchdogs: [Watchdog] (id:2);
-  config: [HmConfig] (id:3);
-}
-
-table Watchdog {
-  shortName: string (id:0);
-  deviceFilePath: string (id:1);
-  maxTimeout: uint32 (id:2);
-  deactivateOnShutdown: bool (id:3);
-  hasValueDeactivateOnShutdown: bool (id:4);
-  requireMagicClose: bool (id:5);
-  hasValueRequireMagicClose: bool (id:6);
-}
-
-table HmConfig {
-  periodicity: uint32 (id:0);
-  bufferSizeMonitor: uint16 (id:1);
-  bufferSizeAliveSupervision: uint16 (id:2);
-}
-
-root_type HMCOREEcuCfg;
