--- score/launch_manager/src/daemon/src/configuration/config_schema/lm_flatcfg.fbs
@@ -1,91 +1,0 @@
-namespace LMFlatBuffer;
-
-// LM version 7.0
-
-file_identifier "BLMR";
-file_extension "bin";
-
-enum ExecutionStateReportingBehaviorEnum : byte {
-  ReportsExecutionState = 0, // REPORTS-EXECUTION-STATE
-  DoesNotReportExecutionState = 1  // DOES-NOT-REPORT-EXECUTION-STATE
-}
-
-enum TerminationBehaviorEnum : byte {
-  ProcessIsNotSelfTerminating = 0, // PROCESS-IS-NOT-SELF-TERMINATING
-  ProcessIsSelfTerminating = 1  // PROCESS-IS-SELF-TERMINATING
-}
-
-table LMEcuCfg {
-  version_major: int (id:0);
-  version_minor: int (id:1);
-  mode_group: [ModeGroup] (id:2);
-  process: [Process] (id:3);
-}
-
-table ModeGroup {
-  identifier: string (id:0);
-  initial_mode_name: string (id:1);
-  initial_mode_value: string (id:2);
-  recovery_mode_name: string (id:3);
-  mode_declaration: [ModeDeclaration] (id:4);
-}
-
-// TODO: Could be moved into ModeGroup
-table ModeDeclaration {
-  identifier: string (id:0);
-}
-
-table Process {
-  identifier: string (id:0);
-  number_of_restart_attempts: uint (id:1);
-  function_cluster_affiliation: string (id:2);
-  coremask: string (id:3);
-  uid: uint (id:4);
-  gid: uint (id:5);
-  path: string (id:6);
-  working_dir: string (id:11);
-  security_policy_details: string (id:7);
-  executable_reporting_behavior: ExecutionStateReportingBehaviorEnum (id:8);
-  startup_config: [ProcessStartupConfig] (id:9);
-  sgids: [ProcessSgid] (id:10);
-}
-
-table ProcessStartupConfig {
-  identifier: string (id:0);
-  enter_timeout_value: uint (id:1);
-  exit_timeout_value: uint (id:2);
-  scheduling_priority: string (id:3);
-  scheduling_policy: string (id:4);
-  execution_error: string (id:5);
-  termination_behavior: TerminationBehaviorEnum (id:6);
-  execution_dependency: [ProcessExecutionDependency] (id:7);
-  process_group_state_dependency: [ProcessGroupStateDependency] (id:8);
-  environment_variable: [EnvironmentVariable] (id:9);
-  process_argument: [ProcessArgument] (id:10);
-  memory_usage: uint64 (id:11);
-}
-
-table ProcessGroupStateDependency {
-  state_machine_name: string (id:0);
-  state_name: string (id:1);
-}
-
-table ProcessExecutionDependency {
-  state_name: string (id:0);
-  target_process_identifier: string (id:1);
-}
-
-table EnvironmentVariable {
-  key: string (id:0);
-  value: string (id:1);
-}
-
-table ProcessArgument {
-  argument: string (id:0);
-}
-
-table ProcessSgid {
-  sgid: uint (id:0);
-}
-
-root_type LMEcuCfg;
