+++ score/launch_manager/src/daemon/src/configuration/details/flatbuffer_config_loader_UT.cpp
@@ -11,7 +11,7 @@
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
 #include "score/mw/launch_manager/configuration/flatbuffer_config_loader.hpp"
-#include "score/launch_manager/src/daemon/src/configuration/details/new_lm_flatcfg_generated.h"
+#include "score/launch_manager/src/daemon/src/configuration/details/lm_flatcfg_generated.h"
 
 #include "score/filesystem/path.h"
 #include "score/os/errno.h"
