+++ score/launch_manager/src/daemon/src/configuration/details/flatbuffer_type_converters.cpp
@@ -13,7 +13,7 @@
 
 #include "score/launch_manager/src/daemon/src/configuration/details/flatbuffer_type_converters.hpp"
 
-#include "score/launch_manager/src/daemon/src/configuration/details/new_lm_flatcfg_generated.h"
+#include "score/launch_manager/src/daemon/src/configuration/details/lm_flatcfg_generated.h"
 
 #include "score/mw/launch_manager/common/log.hpp"
