+++ score/launch_manager/src/daemon/src/configuration/details/flatbuffer_type_converters_UT.cpp
@@ -11,7 +11,7 @@
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
 #include "score/launch_manager/src/daemon/src/configuration/details/flatbuffer_type_converters.hpp"
-#include "score/launch_manager/src/daemon/src/configuration/details/new_lm_flatcfg_generated.h"
+#include "score/launch_manager/src/daemon/src/configuration/details/lm_flatcfg_generated.h"
 
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
