+++ score/launch_manager/src/daemon/src/configuration/details/lm_flatcfg.fbs
